+++ test/autofix/tslintFix.test.ts
@@ -14,23 +14,25 @@
  * limitations under the License.
  */
 
+import * as assert from "power-assert";
+
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
 import { InMemoryFile } from "@atomist/automation-client/project/mem/InMemoryFile";
 
 import { successOn } from "@atomist/automation-client/action/ActionResult";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import { GitCommandGitProject } from "@atomist/automation-client/project/git/GitCommandGitProject";
-import { executeAutofixes } from "@atomist/sdm/api-helper/listener/executeAutofixes";
-import { SingleProjectLoader } from "@atomist/sdm/api-helper/test/SingleProjectLoader";
-import * as assert from "assert";
 import { DefaultRepoRefResolver } from "@atomist/sdm-core";
+import { executeAutofixes } from "@atomist/sdm/api-helper/listener/executeAutofixes";
 import { fakeGoalInvocation } from "@atomist/sdm/api-helper/test/fakeGoalInvocation";
-import { tslintFix } from "../../../src";
+import { SingleProjectLoader } from "@atomist/sdm/api-helper/test/SingleProjectLoader";
+
+import { tslintFix } from "../../lib/autofix/tslintFix";
 
 describe("tsLintFix", () => {
 
     it("should lint and make fixes", async () => {
-        const p = await GitCommandGitProject.cloned({token: null}, new GitHubRepoRef("atomist", "tree-path-ts"));
+        const p = await GitCommandGitProject.cloned({ token: null }, new GitHubRepoRef("atomist", "tree-path-ts"));
         // Make commit and push harmless
         p.commit = async () => {
             return successOn(p);
@@ -43,7 +45,8 @@ describe("tsLintFix", () => {
         // Now mess it up with a lint error
         await p.addFile(f.path, f.content);
 
-        await executeAutofixes(pl, [tslintFix], new DefaultRepoRefResolver())(fakeGoalInvocation(p.id as RemoteRepoRef));
+        await executeAutofixes(pl, [tslintFix],
+            new DefaultRepoRefResolver())(fakeGoalInvocation(p.id as RemoteRepoRef));
         const fileNow = p.findFileSync(f.path);
         assert(!!fileNow, "Did not find file: " + f.path);
         const contentNow = fileNow.getContentSync();