+++ test/fingerprint/PackageLockFingerprinter.test.ts
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
+import * as assert from "power-assert";
+
 import { Fingerprint } from "@atomist/automation-client/project/fingerprint/Fingerprint";
 import { InMemoryFile } from "@atomist/automation-client/project/mem/InMemoryFile";
 import { InMemoryProject } from "@atomist/automation-client/project/mem/InMemoryProject";
 import { PushImpactListenerInvocation } from "@atomist/sdm/api/listener/PushImpactListener";
-import * as assert from "assert";
-import { PackageLockFingerprinter } from "../../../src/support/fingerprint/PackageLockFingerprinter";
+
+import { PackageLockFingerprinter } from "../../lib/fingerprint/PackageLockFingerprinter";
 
 describe("package-lock.json", () => {
 
@@ -27,37 +29,37 @@ describe("package-lock.json", () => {
 
     it("should produce no fingerprint when no package-lock.json", async () => {
         const project = InMemoryProject.of();
-        const cri = {project} as any as PushImpactListenerInvocation;
+        const cri = { project } as any as PushImpactListenerInvocation;
         const fp = await fingerprinter.action(cri);
         assert.equal((fp as any[]).length, 0);
     });
 
     it("should produce fingerprint when package-lock.json found", async () => {
         const project = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1()));
-        const cri = {project} as any as PushImpactListenerInvocation;
+        const cri = { project } as any as PushImpactListenerInvocation;
         const fp = await fingerprinter.action(cri) as Fingerprint;
         assert.equal(fp.abbreviation, "deps");
         assert(!!fp.sha);
     });
 
     it("should not detect change unless dependencies change", async () => {
-        const project1 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({ lockfileVersion: 1})));
-        const cri1 = {project: project1} as any as PushImpactListenerInvocation;
+        const project1 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({ lockfileVersion: 1 })));
+        const cri1 = { project: project1 } as any as PushImpactListenerInvocation;
         const fp1 = await fingerprinter.action(cri1) as Fingerprint;
 
-        const project2 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({lockfileVersion: 2})));
-        const cri2 = {project: project2} as any as PushImpactListenerInvocation;
+        const project2 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({ lockfileVersion: 2 })));
+        const cri2 = { project: project2 } as any as PushImpactListenerInvocation;
         const fp2 = await fingerprinter.action(cri2) as Fingerprint;
         assert.equal(fp1.sha, fp2.sha);
     });
 
     it("should detect change when dependencies change", async () => {
-        const project1 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({antlrVersion: "0.2.0"})));
-        const cri1 = {project: project1} as any as PushImpactListenerInvocation;
+        const project1 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({ antlrVersion: "0.2.0" })));
+        const cri1 = { project: project1 } as any as PushImpactListenerInvocation;
         const fp1 = await fingerprinter.action(cri1) as Fingerprint;
 
-        const project2 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({antlrVersion: "0.2.1"})));
-        const cri2 = {project: project2} as any as PushImpactListenerInvocation;
+        const project2 = InMemoryProject.of(new InMemoryFile("package-lock.json", valid1({ antlrVersion: "0.2.1" })));
+        const cri2 = { project: project2 } as any as PushImpactListenerInvocation;
         const fp2 = await fingerprinter.action(cri2) as Fingerprint;
         assert.notEqual(fp1.sha, fp2.sha);
     });
@@ -66,7 +68,7 @@ describe("package-lock.json", () => {
 
 // tslint:disable
 
-function valid1(params: Partial<{lockfileVersion: number, antlrVersion: string}> = {}) {
+function valid1(params: Partial<{ lockfileVersion: number, antlrVersion: string }> = {}) {
     const paramsToUse = {
         lockfileVersion: 1,
         antlrVersion: "0.2.0",