+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/atomist/sdm-pack-node/compare/0.1.0...HEAD)
 
-## [0.1.0](https://github.com/atomist/sdm-pack-node/tree/0.1.0) - 2018-08-03
+### Changed
+
+-   **BREAKING** Reorganize package to have more standard Node.js
+    layout.  Only import from the main index.
 
 ### Fixed
 
 -   Fix automation client push test. [#2](https://github.com/atomist/sdm-pack-node/issues/2)
+
+## [0.1.0](https://github.com/atomist/sdm-pack-node/tree/0.1.0) - 2018-07-31
+
+Initial release